+++ tfts/models/auto_model.py
@@ -2,6 +2,7 @@
 
 import collections
 import importlib
+import json
 import logging
 import os
 from typing import List, Optional, Tuple, Union
@@ -12,6 +13,9 @@ import tensorflow as tf
 
 from tfts.models.base import BaseConfig, BaseModel
 from tfts.tasks.auto_task import AnomalyHead, ClassificationHead
+
+from ..constants import CONFIG_NAME, TF2_WEIGHTS_INDEX_NAME, TF2_WEIGHTS_NAME, TF_WEIGHTS_NAME
+from .auto_config import AutoConfig
 
 logger = logging.getLogger(__name__)
 
@@ -84,6 +88,36 @@ class AutoModel(BaseModel):
         module = importlib.import_module(f".{model_name}", "tfts.models")
         model = getattr(module, class_name)(config=config, predict_sequence_length=predict_sequence_length)
         return cls(model, config)
+
+    @classmethod
+    def from_pretrained(cls, weights_dir: Union[str, os.PathLike], predict_sequence_length: int = 1):
+        config_path = os.path.join(weights_dir, "config.json")
+        if not os.path.exists(config_path):
+            raise FileNotFoundError(f"Config file not found at {config_path}")
+        try:
+            with open(config_path, "r") as f:
+                config_dict = json.load(f)
+        except Exception as e:
+            raise OSError(f"Error loading config file from {config_path}. Original error: {e}")
+
+        try:
+            model_type = config_dict.get("model_type")
+            if model_type is None:
+                raise ValueError("Missing `model_type` in config.")
+
+            # Dynamically get the correct Config subclass
+            config = AutoConfig.for_model(model_type)
+            config.update(config_dict)  # update with the saved values
+
+            # Build model and load weights
+            model = cls.from_config(config, predict_sequence_length=predict_sequence_length)
+            model.model.load_weights(os.path.join(weights_dir, TF2_WEIGHTS_NAME))
+            return cls(model, config)
+        except Exception as e:
+            raise OSError(
+                f"Error loading model weights from {weights_dir}. "
+                f"Ensure weights were saved using model.save_weights(...). Original error: {e}"
+            )
 
 
 class AutoModelForPrediction(AutoModel):
